Move type and interface
