+++ django_project/frontend/src/components/Home/index.tsx
@@ -21,6 +21,7 @@ import ImageIcon from '@mui/icons-material/Image';
 import {SearchInput} from "../Input/IconInput";
 import {MultipleSelectWithSearch, SelectWithSearch} from "../../components/Input/SelectWithSearch";
 import {SortAscIcon, SortDescIcon} from "../../components/Icons";
+import {GeoSightProject} from '../../types';
 
 import './style.scss';
 
@@ -29,30 +30,6 @@ interface ProjectListProps {
   baseUrl: string;
   setIsLoading: (val: boolean) => void;
 }
-
-type Permission = {
-  list: boolean;
-  read: boolean;
-  edit: boolean;
-  share: boolean;
-  delete: boolean;
-}
-
-export interface GeoSightProject {
-  id: string;               // Unique identifier for the project
-  slug: string;             // Slug for the project
-  icon: string | null;      // Icon for the project (can be null)
-  name: string;             // Name of the project
-  created_at: string;       // Creation timestamp
-  modified_at: string;      // Modification timestamp
-  description: string;      // Description of the project
-  group: string;            // Group associated with the project
-  category: string;         // Category of the project
-  permission: Permission;   // Permissions object
-  reference_layer: number;  // ID of the reference layer
-  creator: number;          // ID of the creator
-}
-
 
 interface ProjectGridProps {
     projects: GeoSightProject[];
@@ -63,12 +40,6 @@ interface ProjectGridProps {
 
 /** Project Grid */
 function ProjectGrid({ projects, onScrollY, containerRef }: ProjectGridProps) {
-  // let newProjects = [];
-  // for (let i = 0; i < 13; i++) {
-  //     let baseProject = projects[i % projects.length];
-  //     if (baseProject) newProjects.push(baseProject);
-  // }
-  // projects = newProjects;
   // @ts-ignore
   const userId: number = user.id;
   return <Grid container spacing={2} className='project-grid-container' onScroll={onScrollY} ref={containerRef}>
